+++ config.py
@@ -1,10 +1,10 @@
 # Single-sourced project configuration values:
 
 # the full version, including alpha/beta/rc tags
-release = '0.7.3'
+release = '0.8.0'
 
 # the short X.Y version
-version = '0.7'
+version = '0.8'
 
 # distribution package name
 name = 'opyprint'